Renamed model class names
 - Kept existing class names for backwards compatibility.
