+++ src/Captcha/Basic.php
@@ -9,7 +9,7 @@ class Basic extends CaptchaAbstract
 {
     public function __construct()
     {
-        if (!config('blogetc.captcha.basic_question') || !config('blogetc.captcha.basic_answers')) {
+        if (! config('blogetc.captcha.basic_question') || ! config('blogetc.captcha.basic_answers')) {
             throw new \DomainException('Invalid question or answers for captcha');
         }
     }
@@ -49,7 +49,7 @@ class Basic extends CaptchaAbstract
             $value = strtolower(trim($value));
             $answers = strtolower($answers);
             $answers_array = array_map('trim', explode(',', $answers));
-            if (!in_array($value, $answers_array, true)) {
+            if (! in_array($value, $answers_array, true)) {
                 return $fail('The captcha field is incorrect.');
             }
         };
