+++ src/Middleware/UserCanManageBlogPosts.php
@@ -18,10 +18,10 @@ class UserCanManageBlogPosts
      */
     public function handle($request, Closure $next)
     {
-        if (!\Auth::check()) {
+        if (! \Auth::check()) {
             abort(401, 'User not authorised to manage blog posts: You are not logged in');
         }
-        if (!\Auth::user()->canManageBlogEtcPosts()) {
+        if (! \Auth::user()->canManageBlogEtcPosts()) {
             abort(401, 'User not authorised to manage blog posts: Your account is not authorised to edit blog posts');
         }
